Approach 1 - MBTI personalities
This was the very first approach I tried to resolve the problem. Initially I believed that applicants could be judged and categorised into smaller subsets based on their personality types. I believed that personality types are important because different industries might have different requirements. For instance, if I as an employer had to hire a door-to-door sales person, I would prefer someone who would be an extrovert over an introvert, someone that would be a risk-taker over someone that would just do what they are told. However, if I were hiring for an IT role, I would not care if a software developer was an introvert or an extrovert.

For categorising all of this, we would first need information about what MBTI personality type's digital footprints actually look liked from already available data on Kaggle, through choice of words and tone analysis. Once I had learned the pattern, I could apply it to the new applicants.

However, upon completion, we went forward with personas rather than MBTI segmentation.